pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "fitz>=0.0.1.dev2",
    "opencv-python-headless>=4.10.0.84",
    "pandas>=2.2.0",
    "pdf2image>=1.17.0",
    "pillow>=10.0.0",
    "psycopg2-binary>=2.9.10",
    "pymupdf>=1.24.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.10",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.3",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
]
